refactor(store): 按域拆分 common/mappers (809→barrel+7 域文件)
- mappers.ts 809→29(barrel)，按 session/world/runtime/state/plugin/memory/snapshot 拆到 common/mappers/
- 各域文件 <210 行，与 records/ 域划分一致，导入路径/符号零变更
- 527 contract 测试全过
